#77 [FEAT] : 따라하기 서비스 분리
